pidmap: introduce pidmap with loadable KO builtable
Introduce pidmap module which can be built as loadable kernel
module.

pidmap is a thread name recorder to provide pid and thread name
mappings which helps debugging. Whenever a thread is created or
renamed, its name is updated in the map in pidmap module.

Note that more memory will be required for this feature, so please
consider the trade-off.

Change-Id: I0cd740bc7e7b7c33b9123f735ca855a8b803456b
Signed-off-by: Stanley Chu <stanley.chu@mediatek.com>
CR-Id: ALPS04362512
Feature: Android Exception Engine(AEE)